Integrate sqldelight

+++ composeApp/src/commonMain/kotlin/mende273/quoteskmp/domain/model/Quote.kt
@@ -3,4 +3,4 @@ package mende273.quoteskmp.domain.model
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class Quote(val id: Int, val content: String, val author: String)
+data class Quote(val id: Long, val content: String, val author: String)
